+++ runzero/api/admin/__init__.py
@@ -3,10 +3,10 @@ admin holds API interactions which require 'account' level authorization. These 
 broad effects across a customer estate and, as such, are kept separately from other API.
 """
 
-from .asset_data_sources import CustomSourcesAdmin
+from .custom_integrations import CustomIntegrationsAdmin
 from .orgs import OrgsAdmin
 
 __all__ = [
-    "CustomSourcesAdmin",
+    "CustomIntegrationsAdmin",
     "OrgsAdmin",
 ]