+++ runzero/api/imports/assets.py
@@ -1,5 +1,5 @@
 """
-enables management of runZero asset data sources.
+enables management of runZero custom integrations.
 
 These operations are privileged and require an account token directly or an OAuth key that can generate one.
 """
@@ -7,7 +7,7 @@ import gzip
 import tempfile
 import time
 import uuid
-from typing import IO, Iterable, List, Optional
+from typing import Iterable, List, Optional
 
 from runzero.client import Client
 from runzero.types import ImportAsset, ImportTask, NewAssetImport, Task
@@ -16,15 +16,15 @@ from runzero.types import ImportAsset, ImportTask, NewAssetImport, Task
 class CustomAssets:
     """Management of Custom Asset Data for your own custom integrations.
 
-    Custom data sources are descriptive registered associations between sources of data
-    and assets imported which are associated with those sources.
+    Custom data integrations are descriptive registered associations between integrations of data
+    and assets imported which are associated with those integrations.
 
     The data sent to the server has basic checks performed and is loaded as an import task
     when it can find the next available worker to do so. Therefore, the result
     is a `class:runzero.Task` which you can check the status of.
 
-    See related :class:`runzero.account.CustomSources` to work with the custom asset data
-    sources registered to the account.
+    See related :class:`runzero.account.CustomIntegrations` to work with the custom asset data
+    integrations registered to the account.
 
     :param client: A handle to the :class:`runzero.Client` which manages interactions
         with the runZero server.
@@ -40,7 +40,7 @@ class CustomAssets:
         self,
         org_id: uuid.UUID,
         site_id: uuid.UUID,
-        source_id: uuid.UUID,
+        custom_integration_id: uuid.UUID,
         assets: List[ImportAsset],
         task_info: Optional[ImportTask] = None,
     ) -> Task:
@@ -49,7 +49,7 @@ class CustomAssets:
 
         :param org_id: Organization ID to import these assets into
         :param site_id: ID of the Site to import these asstes into
-        :param source_id: Custom asset source id for the provided Import Assets
+        :param custom_integration_id: custom integration id for the provided Import Assets
         :param assets: A collection of ImportAssets to upload
         :param task_info: Descriptive information associated with the import
             task to be created. If omitted, a task name is generated for you
@@ -69,7 +69,7 @@ class CustomAssets:
 
         asset_import_req = _create_custom_asset_request(
             site_id=site_id,
-            source_id=source_id,
+            custom_integration_id=custom_integration_id,
             import_task=task_info,
             assets=assets,
         )
@@ -80,7 +80,7 @@ class CustomAssets:
         multipart_form_data = (
             ("assetData", ("asset_data.jsonl.gz", asset_import_req.asset_data)),
             ("siteId", (None, str(asset_import_req.site_id))),
-            ("sourceId", (None, str(asset_import_req.source_id))),
+            ("customIntegrationId", (None, str(asset_import_req.custom_integration_id))),
             ("importTask.name", (None, asset_import_req.import_task.name)),
             ("importTask.description", (None, asset_import_req.import_task.description)),
             ("importTask.tags", (None, tags_as_str)),
@@ -89,24 +89,24 @@ class CustomAssets:
         return Task.parse_obj(res.json_obj)
 
 
-def _import_assets_into_gzip_jsonl(import_assets: Iterable[ImportAsset]) -> IO[bytes]:
+def _import_assets_into_gzip_jsonl(import_assets: Iterable[ImportAsset]) -> bytes:
     tmp = tempfile.TemporaryFile(mode="w+b")
     with gzip.GzipFile(fileobj=tmp, mode="wb") as gzw:
         for asset_obj in import_assets:
             gzw.write(asset_obj.json(by_alias=True).encode("utf-8") + "\n".encode("utf-8"))
     tmp.seek(0)
-    return tmp
+    return tmp.read()
 
 
 def _create_custom_asset_request(
-    site_id: uuid.UUID, source_id: uuid.UUID, assets: Iterable[ImportAsset], import_task: ImportTask
+    site_id: uuid.UUID, custom_integration_id: uuid.UUID, assets: Iterable[ImportAsset], import_task: ImportTask
 ) -> NewAssetImport:
     # TODO: We are disabling validation on all fields with .construct
     # until openapi 'bytes' type and pydantic can agree on a file-like.
     # See FastAPI implementation of UploadFile for ref
-    return NewAssetImport.construct(
+    return NewAssetImport(
         siteId=site_id,
-        sourceId=source_id,
+        customIntegrationId=custom_integration_id,
         importTask=import_task,
-        assetData=_import_assets_into_gzip_jsonl(assets),  # type: ignore
+        assetData=_import_assets_into_gzip_jsonl(assets),
     )