+++ runzero/client/__init__.py
@@ -4,11 +4,14 @@ The Client is responsible for communication with runZero services.
 """
 
 from runzero.client.client import Client
-from runzero.client.errors import AuthError, ClientError, ServerError
+from runzero.client.errors import AuthError, ClientError, RateLimitError, ServerError
+from runzero.types import RateLimitInformation
 
 __all__ = [
     "AuthError",
     "Client",
     "ClientError",
+    "RateLimitError",
+    "RateLimitInformation",
     "ServerError",
 ]